RF: Change behaviour of reference links when reference is missing.
To enhance the integration with Dokuwiki, a reference link that is
missing a reference definition is taken to be link to an internal
wiki page. For example, the links "[somepage]", "[somepage][]", or
"[somepage link text][somepage]" that have a corresponding
"[somepage]: url" will produce a link to "doku.php?id=somepage".